webshop fund0request cleanup refactoring
