+++ app/Http/Controllers/Api/Platform/FundRequests/FundRequestClarificationsController.php
@@ -1,33 +1,30 @@
 <?php
 
-namespace App\Http\Controllers\Api\Platform\Funds\Requests;
+namespace App\Http\Controllers\Api\Platform\FundRequests;
 
-use App\Http\Requests\Api\Platform\Funds\Requests\Clarifications\UpdateFundRequestClarificationsRequest;
+use App\Http\Controllers\Controller;
+use App\Http\Requests\Api\Platform\FundRequests\FundRequestClarifications\UpdateFundRequestClarificationRequest;
 use App\Http\Resources\FundRequestClarificationResource;
-use App\Models\Fund;
 use App\Models\FundRequest;
 use App\Models\FundRequestClarification;
-use App\Http\Controllers\Controller;
 
 class FundRequestClarificationsController extends Controller
 {
     /**
      * Update the specified resource in storage.
      *
-     * @param UpdateFundRequestClarificationsRequest $request
-     * @param Fund $fund
+     * @param UpdateFundRequestClarificationRequest $request
      * @param FundRequest $fundRequest
      * @param FundRequestClarification $requestClarification
      * @return FundRequestClarificationResource
      * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function update(
-        UpdateFundRequestClarificationsRequest $request,
-        Fund $fund,
+        UpdateFundRequestClarificationRequest $request,
         FundRequest $fundRequest,
         FundRequestClarification $requestClarification
     ): FundRequestClarificationResource {
-        $this->authorize('update', [$requestClarification, $fundRequest, $fund]);
+        $this->authorize('update', [$requestClarification, $fundRequest]);
 
         $requestClarification->update(array_merge($request->only('answer'), [
             'answered_at' => now(),
