+++ routes/api-platform.php
@@ -53,17 +53,6 @@ $router->group([], static function() use ($router) {
                 'records' => 'fund_request_record',
             ]
         ]);
-
-        $router->resource(
-            'funds/{fund}/requests/{fund_request}/clarifications',
-            "Api\Platform\Funds\Requests\FundRequestClarificationsController", [
-            'only' => [
-                'update'
-            ],
-            'parameters' => [
-                'clarifications' => 'fund_request_clarification',
-            ]
-        ]);
     }
 
     $router->resource(
@@ -520,6 +509,13 @@ $router->group(['middleware' => 'api.auth'], static function() use ($router) {
         $router->resource('fund-requests', "Api\Platform\FundRequestsController")->only([
             'index', 'show',
         ]);
+
+        $router->resource(
+            'fund-requests/{fund_request}/clarifications',
+            'Api\Platform\FundRequests\FundRequestClarificationsController'
+        )->parameters([
+            'clarifications' => 'fund_request_clarification',
+        ])->only('update');
     }
 
     $router->get(
